Refactor RoomFooterViewModel
