+++ app-android/app/src/main/kotlin/com/example/home_hackathon2/di/ViewModelEntryPoint.kt
@@ -2,7 +2,7 @@ package com.example.home_hackathon2.di
 
 import com.example.home_hackathon2.ui.AppViewModel
 import com.example.home_hackathon2.ui.chat.ChatRoomViewModel
-import com.example.home_hackathon2.ui.chat.MicInViewModel
+import com.example.home_hackathon2.ui.chat.RoomFooterViewModel
 import com.example.home_hackathon2.ui.initial.InitialViewModel
 import dagger.hilt.EntryPoint
 import dagger.hilt.InstallIn
@@ -14,5 +14,5 @@ interface ViewModelEntryPoint {
     fun getAppViewModel(): AppViewModel
     fun getInitialViewModel(): InitialViewModel
     fun getChatRoomViewModel(): ChatRoomViewModel
-    fun getMicInViewModel(): MicInViewModel
+    fun getRoomFooterViewModel(): RoomFooterViewModel
 }
