+++ app-android/app/src/main/kotlin/com/example/home_hackathon2/ui/chat/ChatRoomScreen.kt
@@ -1,22 +1,12 @@
 package com.example.home_hackathon2.ui.chat
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import android.content.Intent
-import android.speech.RecognizerIntent
-import android.speech.SpeechRecognizer
 import androidx.compose.animation.Crossfade
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.*
-import androidx.compose.material.ripple.rememberRipple
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
@@ -24,20 +14,16 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.example.home_hackathon2.R
 import com.example.home_hackathon2.model.ChatRoom
-import com.example.home_hackathon2.ui.listener.SimpleRecognizerListener
 import com.example.home_hackathon2.ui.res.*
 import com.example.home_hackathon2.ui.tools.rememberViewModel
 import com.example.home_hackathon2.ui.widget.paddingInsets
 
 @Composable
 fun ChatRoomScreen() {
-    val chatRoomViewModel = rememberViewModel {
+    val viewModel = rememberViewModel {
         it.getChatRoomViewModel()
     }
-    val micInViewModel = rememberViewModel {
-        it.getMicInViewModel()
-    }
-    val chatRoom = chatRoomViewModel.chatRoom.collectAsState()
+    val chatRoom = viewModel.chatRoom.collectAsState()
     Box(
         modifier = Modifier
             .paddingInsets(top = true, bottom = true)
@@ -51,14 +37,9 @@ fun ChatRoomScreen() {
                 chatRoom = chatRoom.value,
                 modifier = Modifier.weight(1F)
             )
-            Footer()
+            RoomFooter()
         }
     }
-    SpeechRecognizer(
-        onRecognized = {
-            micInViewModel.send(it)
-        }
-    )
 }
 
 @Composable
@@ -106,71 +87,6 @@ private fun Header() {
 
 @Preview
 @Composable
-private fun Footer() {
-    Box(
-        modifier = Modifier.height(56.dp)
-    ) {
-        Box(
-            modifier = Modifier
-                .offset(y = (-30).dp)
-                .requiredWidth(62.dp)
-                .requiredHeight(62.dp)
-                .clip(CircleShape)
-                .background(COLOR_LIGHT)
-                .align(Alignment.TopCenter)
-        )
-        Spacer(
-            modifier = Modifier
-                .fillMaxWidth()
-                .align(Alignment.TopCenter)
-                .height(1.dp)
-                .background(COLOR_LIGHT)
-        )
-        MicButton(
-            modifier = Modifier
-                .offset(y = (-30).dp)
-                .requiredWidth(60.dp)
-                .requiredHeight(60.dp)
-                .align(Alignment.TopCenter)
-        )
-    }
-}
-
-@Preview
-@Composable
-private fun MicButton(
-    modifier: Modifier = Modifier
-) {
-    Box(
-        modifier = modifier
-            .clip(CircleShape)
-            .width(60.dp)
-            .height(60.dp)
-            .background(COLOR_WHITE)
-            .clickable(
-                interactionSource = remember { MutableInteractionSource() },
-                indication = rememberRipple(),
-                onClick = {}
-            )
-            .border(2.dp, COLOR_PRIMARY, CircleShape)
-    ) {
-        Icon(
-            modifier = Modifier
-                .clip(CircleShape)
-                .width(52.dp)
-                .height(52.dp)
-                .background(color = COLOR_PRIMARY)
-                .padding(10.dp)
-                .align(Alignment.Center),
-            painter = painterResource(id = R.drawable.ic_microphone_solid),
-            contentDescription = null,
-            tint = COLOR_WHITE
-        )
-    }
-}
-
-@Preview
-@Composable
 private fun Empty() {
     Column(
         modifier = Modifier.fillMaxSize(),
@@ -190,45 +106,6 @@ private fun Empty() {
             color = COLOR_DARK,
             textAlign = TextAlign.Center
         )
-    }
-}
-
-
-@Composable
-fun SpeechRecognizer(
-    onRecognized: (String) -> Unit
-) {
-    val context = LocalContext.current
-    val speechRecognizer = remember {
-        SpeechRecognizer.createSpeechRecognizer(context)
-    }
-    val intent = remember {
-        Intent(RecognizerIntent.ACTION_RECOGNIZE_SPEECH)
-    }
-    intent.putExtra(
-        RecognizerIntent.EXTRA_LANGUAGE_MODEL,
-        RecognizerIntent.LANGUAGE_MODEL_FREE_FORM
-    )
-    intent.putExtra(RecognizerIntent.EXTRA_CALLING_PACKAGE, LocalContext.current.packageName)
-
-    val callback = remember {
-        object : SimpleRecognizerListener.SimpleRecognizerResponseListener {
-            override fun onRecognizedResult(speechText: String) {
-                if (speechText.isNotBlank()) {
-                    onRecognized(speechText)
-                }
-                speechRecognizer.stopListening()
-                speechRecognizer.startListening(intent)
-            }
-        }
-    }
-
-    DisposableEffect(true) {
-        speechRecognizer.setRecognitionListener(SimpleRecognizerListener(callback))
-        speechRecognizer.startListening(intent)
-        onDispose {
-            speechRecognizer.stopListening()
-        }
     }
 }
